Integer.toString to Double.toString

+++ src/java/relex/parser/RemoteLGParser.java
@@ -139,8 +139,8 @@ public class RemoteLGParser extends LGParser
 			// set meta data
 			FeatureNode meta = new FeatureNode();
 			meta.set("num_skipped_words", new FeatureNode(Integer.toString(parseResult.getNumSkippedWords())));
-			meta.set("disjunct_cost", new FeatureNode(Integer.toString(linkage.getDisjunctCost())));
-			meta.set("link_cost", new FeatureNode(Integer.toString(linkage.getLinkCost())));
+			meta.set("disjunct_cost", new FeatureNode(Double.toString(linkage.getDisjunctCost())));
+			meta.set("link_cost", new FeatureNode(Double.toString(linkage.getLinkCost())));
 			meta.set("num_violations", new FeatureNode(Integer.toString(linkage.getNumViolations())));
 			parsedSentence.setMetaData(meta);
 		}
